pyproject.toml
[tool.poetry]
name = "financial-reports"
version = "0.1.0"
description = "Gather financial reports from different sources"
authors = ["Wayne <waynexu1027@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = {extras = ["http2"], version = "^0.27.2"}
pydantic = "^2.9.2"
pydantic-settings = "^2.6.1"
sqlmodel = "^0.0.22"
click = "^8.1.7"
google-api-python-client = "^2.154.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.4"
pre-commit = "^4.0.1"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191", # indentation contains tabs
]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# mypy type checking
# https://docs.pydantic.dev/latest/integrations/mypy/
[tool.mypy]
plugins = [
  "pydantic.mypy"
]
pretty = true
ignore_missing_imports = true
show_error_context = true
follow_imports = "skip"
strict_optional = true
show_column_numbers = true
disallow_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
disable_error_code = [
    "import-untyped",
    "attr-defined"
]
enable_incomplete_feature = ["NewGenericSyntax"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
